Compatibility with ARC
This commit replaces the DDGetoptOption struct with an NSObject to add support for ARC. For easy transition of existing code, just change `DDGetoptOption` to `OldDDGetoptOption`. The strings will be copied in `-addOptionsFromTable:`, but note that the table itself will become invalid when the autorelease pool is drained, since the table can't retain the long option strings.